Move Store to client package
