+++ protoarray/src/test/java/tech/pegasys/teku/protoarray/FFGUpdatesTest.java
@@ -27,8 +27,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
-import tech.pegasys.teku.storage.Store;
 import tech.pegasys.teku.storage.api.StubStorageUpdateChannel;
+import tech.pegasys.teku.storage.client.Store;
 
 public class FFGUpdatesTest {
