+++ protoarray/src/test/java/tech/pegasys/teku/protoarray/NoVotesTest.java
@@ -28,8 +28,8 @@ import java.util.Collections;
 import java.util.List;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
-import tech.pegasys.teku.storage.Store;
 import tech.pegasys.teku.storage.api.StubStorageUpdateChannel;
+import tech.pegasys.teku.storage.client.Store;
 
 public class NoVotesTest {
