+++ protoarray/src/test/java/tech/pegasys/teku/protoarray/ProtoArrayForkChoiceStrategyTest.java
@@ -31,8 +31,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.datastructures.forkchoice.MutableStore;
 import tech.pegasys.teku.datastructures.forkchoice.VoteTracker;
-import tech.pegasys.teku.storage.Store;
 import tech.pegasys.teku.storage.api.StubStorageUpdateChannel;
+import tech.pegasys.teku.storage.client.Store;
 
 public class ProtoArrayForkChoiceStrategyTest {
