+++ storage/src/test/java/tech/pegasys/teku/storage/client/StoreTest.java
@@ -28,9 +28,8 @@ import tech.pegasys.teku.core.ChainBuilder;
 import tech.pegasys.teku.core.StateTransitionException;
 import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.storage.Store;
-import tech.pegasys.teku.storage.Store.Transaction;
 import tech.pegasys.teku.storage.api.StorageUpdateChannel;
+import tech.pegasys.teku.storage.client.Store.Transaction;
 import tech.pegasys.teku.util.async.SafeFuture;
 
 class StoreTest {
