+++ storage/src/test/java/tech/pegasys/teku/storage/server/AbstractStorageBackedDatabaseTest.java
@@ -30,9 +30,9 @@ import tech.pegasys.teku.core.StateTransition;
 import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.storage.Store;
-import tech.pegasys.teku.storage.Store.Transaction;
 import tech.pegasys.teku.storage.api.DatabaseBackedStorageUpdateChannel;
+import tech.pegasys.teku.storage.client.Store;
+import tech.pegasys.teku.storage.client.Store.Transaction;
 import tech.pegasys.teku.util.config.StateStorageMode;
 import tech.pegasys.teku.util.file.FileUtil;
