+++ storage/src/test/java/tech/pegasys/teku/storage/server/rocksdb/V3RocksDbDatabaseTest.java
@@ -30,7 +30,7 @@ import tech.pegasys.teku.core.ChainBuilder;
 import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.storage.Store;
+import tech.pegasys.teku.storage.client.Store;
 import tech.pegasys.teku.storage.server.Database;
 import tech.pegasys.teku.util.config.StateStorageMode;
